% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_json.R
\name{as.json}
\alias{as.json}
\title{Coerce string to JSON}
\usage{
as.json(x)
}
\arguments{
\item{x}{string to coerce to JSON}
}
\description{
Coerce string to JSON
}
\examples{

js <- '{"x":1,"y":2}'
as.json(js)

}
