% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty.R
\name{minify_json}
\alias{minify_json}
\title{Minify Json}
\usage{
minify_json(json, ...)
}
\arguments{
\item{json}{string of JSON}

\item{...}{other argments passed to \link{to_json}}
}
\description{
Removes indentiation from a JSON string
}
\examples{

df <- data.frame(id = 1:10, val = rnorm(10))
js <- to_json( df )
jsp <- pretty_json(js)
minify_json( jsp )

}
