% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty.R
\name{pretty_json}
\alias{pretty_json}
\title{Pretty Json}
\usage{
pretty_json(json, ...)
}
\arguments{
\item{json}{string of JSON}

\item{...}{other argments passed to \link{to_json}}
}
\description{
Adds indentiation to a JSON string
}
\examples{

df <- data.frame(id = 1:10, val = rnorm(10))
js <- to_json( df )
pretty_json(js)

## can also use directly on an R object
pretty_json( df )

}
