% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_json.R
\name{to_json.default}
\alias{to_json.default}
\title{#' @export
to_json.list <- function( x ) rcpp_list_to_json( x )}
\usage{
\method{to_json}{default}(x)
}
\description{
#' @export
to_json.list <- function( x ) rcpp_list_to_json( x )
}
