% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_json.R
\name{to_json}
\alias{to_json}
\title{To JSON}
\usage{
to_json(x, numeric_dates = TRUE)
}
\arguments{
\item{x}{object to convert to JSON}

\item{numeric_dates}{logical indicating if dates should be treated as numerics. 
Defaults to TRUE for speed. If FALSE, the dates will be coerced to character}
}
\description{
Converts a data.frame to JSON
}
