% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_json.R
\name{to_ndjson}
\alias{to_ndjson}
\title{To ndjson}
\usage{
to_ndjson(x, unbox = FALSE, digits = NULL, numeric_dates = TRUE,
  factors_as_string = TRUE, by = "row")
}
\description{
Converts R objects to ndjson
}
