% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{validate}
\usage{
validate(json)
}
\arguments{
\item{json}{character or json object}
}
\value{
logical vector
}
\description{
Validates JSON
}
\examples{

validate('[]')
df <- data.frame(id = 1:5, val = letters[1:5])
validate( to_json(df) )

validate('{"x":1,"y":2,"z":"a"}')

validate( c('{"x":1,"y":2,"z":"a"}', to_json(df) ) )
validate( c('{"x":1,"y":2,"z":a}', to_json(df) ) )

}
