% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_json}
\alias{validate_json}
\title{validate JSON}
\usage{
validate_json(json)
}
\arguments{
\item{json}{character or json object}
}
\value{
logical vector
}
\description{
Validates JSON
}
\examples{

validate_json('[]')
df <- data.frame(id = 1:5, val = letters[1:5])
validate_json( to_json(df) )

validate_json('{"x":1,"y":2,"z":"a"}')

validate_json( c('{"x":1,"y":2,"z":"a"}', to_json(df) ) )
validate_json( c('{"x":1,"y":2,"z":a}', to_json(df) ) )

}
