% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_json.R
\name{from_json}
\alias{from_json}
\title{From JSON}
\usage{
from_json(x)
}
\arguments{
\item{x}{JSON string to convert to R list}
}
\description{
Converts a JSON string to an R list
}
\examples{

from_json("{\\"a\\":[1, 2, 3]}")
from_json("{\\"a\\":8, \\"b\\":99.5, \\"c\\":true, \\"d\\":\\"cats\\", \\"e\\":[1, \\"cats\\", 3]}")
from_json("{\\"a\\":8, \\"b\\":{\\"c\\":123, \\"d\\":{\\"e\\":456}}}")

json_str <- jsonify::to_json(list("a" = 5L, "b" = 1.43, "c" = "cats", "d" = FALSE))
from_json(json_str)

}
